KhoHang_API/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
email-validator
google-genai
python-multipart
# Image processing
Pillow
# Export dependencies
openpyxl
reportlab
num2words
# Database
sqlalchemy
PyJWT
passlib[bcrypt]
python-jose[cryptography]
